requirements.txt
torch
apscheduler == 3.6.3
psycopg2-binary>=2.8
tqdm
python-dotenv
notebook
requests
matplotlib